refactor: 성능 최적화를 위한 useMemo 사용
